Move edit to separate file
