+++ client/gutenberg/extensions/related-posts/constants.js
@@ -1,0 +1,1 @@
+export const MAX_POSTS_TO_SHOW = 3;
